requirements.txt
python@3.11
fastapi
uvicorn
sqlalchemy
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
pyspark
requests
redis
ray
pyarrow
getdaft
lancedb